2008-05-01  Diego Novillo  <dnovillo@google.com>
	http://gcc.gnu.org/ml/gcc-patches/2008-05/msg00053.html

	* tree-vrp.c (vrp_visit_phi_node): Cast variable I to int
	for printing.
	* cgraph.c (cgraph_release_function_body): Only call
	gimple_set_body if NODE->DECL has a struct function.
	* tree.c (make_node_stat): Do not call gimple_set_body.
	* cp/Make-lang.in (cp/semantics.o): Add dependency on
	$(GIMPLE_H).
	* cp/semantics.c: Include gimple.h
	(expand_or_defer_fn): Assert that FN has a gimple body.
	* function.h (struct function): Add field GIMPLE_BODY.
	* gimple.c (gimple_bodies_vec): Remove.
	(gimple_bodies_map): Remove.
	(gimple_set_body): Re-write to use GIMPLE_BODY in FN's
	function object.
	(gimple_body): Likewise.

2008-05-01  Oleg Ryjkov  <olegr@google.com>

	http://gcc.gnu.org/ml/gcc-patches/2008-05/msg00053.html

	* tree-eh.c (record_in_goto_queue, record_in_goto_queue_label):
	New functions.
	(maybe_record_in_goto_queue): Refactored and added support for
	recording labels embedded in GIMPLE_CONDs.
	(lower_catch, lower_cleanup): Fix 3 typos that were introduced
	during the conversion to tuples.
	(lower_eh_constructs_2): Call maybe_record_in_goto_queue for
	GIMPLE_CONDs.




git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gimple-tuples-branch@134863 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/tree-vrp.c
@@ -6074,7 +6074,7 @@ vrp_visit_phi_node (gimple phi)
 	{
 	  fprintf (dump_file,
 	      "\n    Argument #%d (%d -> %d %sexecutable)\n",
-	      i, e->src->index, e->dest->index,
+	      (int) i, e->src->index, e->dest->index,
 	      (e->flags & EDGE_EXECUTABLE) ? "" : "not ");
 	}
